refactor(xmtp_mls): fold key_package_cleaner remnants into key_package_maintenance
The worker is gone; a Worker-named struct wrapping two delete helpers was
gutting-stage scaffolding. Helpers become free fns in the maintenance module,
the error is renamed KeyPackageMaintenanceError to match its actual role, and
the file is deleted.
